// src/domain/uhTipo/types.ts
export type UhTipo = {
  id: number;
  iduhtipo: number;
  idempresa: number;
  empresa_dsabreviatura: string;
  cduhtipo: string;
  nmuhtipo: string;
  dsuhtipo: string | null;
  iduhtipogrupo: number | null;
  nmtipogrupo: string | null;
  idpai: number | null;
  qtleito: number;
  fltipocobranca: number;
  flsituacao: number;
};